+++ resources/lang/en/nav.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'menu_aria' => 'Account menu',
+];