+++ src/Http/Middleware/ApplyVpressSiteConfig.php
@@ -7,6 +7,7 @@ namespace Voodflow\Vpress\Http\Middleware;
 use Closure;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Voodflow\Tutorials\Support\LocaleSwitcher;
 use Voodflow\Vpress\Models\VpressSettings;
 
 class ApplyVpressSiteConfig
@@ -16,6 +17,10 @@ class ApplyVpressSiteConfig
         config([
             'seo.canonical_link' => (bool) VpressSettings::get('seo_canonical_enabled', true),
         ]);
+
+        if (class_exists(LocaleSwitcher::class)) {
+            app()->setLocale(VpressSettings::primaryLocale());
+        }
 
         return $next($request);
     }